Fixes issue #5
- Default state of this extension is to allow CSP headers. A user must
click the icon to disable CSP headers. This follows the principle of
least surprise.

- Fix incorrect icon being shown initially. The extension used to show
the greyed 'CSP headers enabled' icon when in fact the extension was
disabling CSP headers. When you clicked the icon, it continued to be
grey. A third click correctly toggled the state of the extension.

- Add a title/tooltip to explain if CSP headers are being disabled or
not.

- Update the extension description to explain the icon colours.